test_complete_system.py
"""
Complete system test for the AI-powered hazard image validation
"""

import requests
import sys
import os
from PIL import Image
import time

def create_test_images():
    """Create test images for validation"""
    
    # 1. Valid hazard image - road damage simulation
    print("Creating valid hazard image...")
    hazard_image = Image.new('RGB', (400, 400), color=(120, 120, 120))
    
    # Add dark patches (potholes)
    for i in range(50, 150):
        for j in range(50, 150):
            hazard_image.putpixel((i, j), (20, 20, 20))
    
    # Add more dark patches
    for i in range(200, 300):
        for j in range(200, 300):
            hazard_image.putpixel((i, j), (15, 15, 15))
    
    hazard_image.save('test_pothole_damage.jpg')
    
    # 2. Invalid image - clear sky
    print("Creating invalid (non-hazard) image...")
    sky_image = Image.new('RGB', (300, 300), color=(135, 206, 235))
    sky_image.save('test_clear_sky.jpg')
    
    # 3. Fire hazard image
    print("Creating fire hazard image...")
    fire_image = Image.new('RGB', (350, 350), color=(220, 100, 50))
    
    # Add some red/orange areas
    for i in range(100, 250):
        for j in range(100, 250):
            fire_image.putpixel((i, j), (255, 69, 0))
    
    fire_image.save('test_fire_emergency.jpg')
    
    return ['test_pothole_damage.jpg', 'test_clear_sky.jpg', 'test_fire_emergency.jpg']

def test_image_validation(image_path):
    """Test image validation via API"""
    try:
        print(f"\nTesting image: {image_path}")
        
        with open(image_path, 'rb') as f:
            files = {'image': f}
            response = requests.post('http://127.0.0.1:5000/hazards/api/validate-image', files=files)
        
        if response.status_code == 200:
            data = response.json()
            print(f"✓ API Response: {data}")
            
            if data['success'] and data['is_hazard']:
                print(f"  ✅ VALID HAZARD - Confidence: {data['confidence']}% | Type: {data['hazard_type']}")
                return True
            else:
                print(f"  ❌ INVALID IMAGE - {data['message']}")
                return False
        else:
            print(f"✗ API Error: {response.status_code} - {response.text}")
            return False
            
    except Exception as e:
        print(f"✗ Error testing {image_path}: {str(e)}")
        return False

def cleanup_test_files(files):
    """Clean up test files"""
    for file in files:
        if os.path.exists(file):
            os.remove(file)
            print(f"Cleaned up: {file}")

def main():
    print("🤖 AI-Powered Hazard Image Validation System Test")
    print("=" * 60)
    
    # Check if server is running
    try:
        response = requests.get('http://127.0.0.1:5000/')
        if response.status_code != 200:
            print("❌ Flask server is not running. Please start it first with: python run.py")
            return False
    except requests.exceptions.ConnectionError:
        print("❌ Flask server is not running. Please start it first with: python run.py")
        return False
    
    print("✅ Flask server is running")
    
    # Create test images
    test_files = create_test_images()
    
    # Test each image
    results = []
    for image_path in test_files:
        result = test_image_validation(image_path)
        results.append(result)
        time.sleep(1)  # Small delay between requests
    
    print(f"\n{'='*60}")
    print("📊 TEST RESULTS SUMMARY:")
    print(f"{'='*60}")
    
    expected_results = [True, False, True]  # pothole: valid, sky: invalid, fire: valid
    
    for i, (file, result, expected) in enumerate(zip(test_files, results, expected_results)):
        status = "✅ PASS" if result == expected else "❌ FAIL"
        print(f"{i+1}. {file}: {status}")
    
    # Overall success rate
    success_rate = sum(1 for r, e in zip(results, expected_results) if r == e) / len(results) * 100
    print(f"\n🎯 Overall Success Rate: {success_rate:.1f}%")
    
    if success_rate >= 66:  # At least 2 out of 3 should work
        print("🏆 AI Image Validation System is working correctly!")
        overall_success = True
    else:
        print("⚠️ AI Image Validation System needs improvement")
        overall_success = False
    
    print(f"\n{'='*60}")
    print("💡 SYSTEM FEATURES DEMONSTRATED:")
    print("✅ Real-time AI image classification")
    print("✅ Hazard vs Non-hazard detection")
    print("✅ Confidence scoring")
    print("✅ Hazard type classification")
    print("✅ REST API integration")
    print("✅ User-friendly error messages")
    
    # Cleanup
    cleanup_test_files(test_files)
    
    return overall_success

if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)